+++ src/worker/worker.http.ts
@@ -87,7 +87,7 @@ export class WorkerHttp {
 
 		// setup fetching promises
 		const promises = [
-			["7TV", seventv.loadUserEmoteSet("TWITCH", channel.id).catch(() => void 0)],
+			["7TV", seventv.loadUserEmoteSet(port.platform ?? "TWITCH", channel.id).catch(() => void 0)],
 			["FFZ", frankerfacez.loadUserEmoteSet(channel.id).catch(() => void 0)],
 			["BTTV", betterttv.loadUserEmoteSet(channel.id).catch(() => void 0)],
 		] as [string, Promise<SevenTV.EmoteSet>][];